Modify 1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -8,7 +8,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	int divisor = 1, temp_num;
+	unsigned long int divisor = 1, temp_num;
 
 	if (n == 0)
 	{
